requirements.txt
discord.py
pyTelegramBotAPI
aiofiles
aiohttp
moviepy
motor